Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int a,b,c;
+    double a,b,c;
     scanf("%d %d %d",&a,&b,&c);
-    printf("Average: %d\n",a+b+c/3);
+    printf("Average: %.2lf\n",a+b+c/3);
     return 0;
 }
